pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "hemera-oneiros"
version = "0.1.0"
description = "Reference scaffolding for the Hemera OneirOS zkVM stack"
readme = "docs/DOCUMENTATION.md"
authors = [{name = "Hemera Labs"}]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[project.scripts]
oneiros = "oneiros.ops.cli:main"

[tool.setuptools.packages.find]
include = ["oneiros*"]
